requirements.txt
streamlit==1.28.0
pandas==2.0.3
Pillow==10.0.0
numpy==1.24.3
matplotlib==3.7.2
altair==5.0.1streamlit==1.31.0
pandas==2.1.0
pillow==10.0.0